deconstructed charity pages routing & business logic to CSR components. Also moved loadingOverlay to loading.tsx in app root
